Refactored util directory to simplify end user view
