+++ src/Synful/Util/DataManagement/SqlConnection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Synful\DataManagement;
+namespace Synful\Util\DataManagement;
 
 use MySqli;
